Keep whoami in one module and open stores observe-only
Collapse the actor/hosted/report split back into whoami.rs. Probe with
discover_heddle_root and open only an already-present store so whoami
does not bootstrap a Git sidecar. Surface open errors on a broken
Heddle repo instead of falling through to user_config. Dedup the
without-repo principal tail. Revert #1392 catalog drift in generated
schemas; keep only WhoamiCaptureActorSchema.

Fixes #1437
